// SDKs/twilio.com/twilio_chat_v1/1.36.0/typescript/src/sdk/models/shared/chatv1serviceuser.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class ChatV1ServiceUser extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=attributes" })
  attributes?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=identity" })
  identity?: string;

  @Metadata({ data: "json, name=is_notifiable" })
  isNotifiable?: boolean;

  @Metadata({ data: "json, name=is_online" })
  isOnline?: boolean;

  @Metadata({ data: "json, name=joined_channels_count" })
  joinedChannelsCount?: number;

  @Metadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @Metadata({ data: "json, name=role_sid" })
  roleSid?: string;

  @Metadata({ data: "json, name=service_sid" })
  serviceSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}